+++ styles/components/PokemonCard.style.ts
